tester: differentiate expected output by golang version
Fixes #3082.
